Project skeleton/general structure created.
Created multiple folders and files that will be used when creating the rest of the project, will be following this general structure but it may change in future commits depending on situation and context. Some of these files already have code in them, but all code dealing with the SQLAlchemy database have been commented out as I still need to figure out the structure needed for the tables and columns. Will fix this later.